backend/requirements.txt
fastapi==0.115.12
openai==1.82.0
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==1.0.0
uvicorn==0.34.2
